Measure max temperature sample

+++ sw/Drivers/MLX90614/mlx90614_driver.h
@@ -56,9 +56,9 @@
 #define MLX90614_DBG_MSG_R 1
 
 
-int32_t mlx90614GetTemp(I2C_HandleTypeDef * i2cHandle, uint8_t data_addr, uint16_t * data);
-int32_t mlx90614GetObjectTemp(I2C_HandleTypeDef * i2cHandle, float * data);
-int32_t mlx90614GetEnviromentTemp(I2C_HandleTypeDef * i2cHandle, float * data);
+int16_t mlx90614GetTemp(I2C_HandleTypeDef * i2cHandle, uint8_t data_addr, uint16_t * data);
+int16_t mlx90614GetObjectTemp(I2C_HandleTypeDef * i2cHandle, float * data);
+int16_t mlx90614GetEnviromentTemp(I2C_HandleTypeDef * i2cHandle, float * data);
 int32_t mlx90632_i2c_write(I2C_HandleTypeDef * i2cHandle, int16_t register_address, uint16_t value);
 int32_t mlx90632_i2c_read32(I2C_HandleTypeDef * i2cHandle, int16_t register_address, uint32_t *value);
 int32_t mlx90632_i2c_read(I2C_HandleTypeDef * i2cHandle, int16_t register_address, uint16_t *value);
